tests: move init_authsets_ fixtures to common
will be reused by preauthorization tests

changelog: none
